refactor(src-tauri/tts): move map_via_spans out of the piper module
The normalized-to-original char-offset mapping is engine-agnostic (it
mirrors ttsd's _map_via_spans), yet it lived in tts::piper::timestamps
and the Silero Native engine imported it from across the piper module
boundary. Move it — with its tests — to tts/mod.rs next to
CharMappingEntry.
